refactor(agent): split mcp_server.rs into integrations/mcp submodule
Converts apps/agent/src-tauri/src/integrations/mcp_server.rs (2,022 LOC) into
integrations/mcp/ with separate files for server lifecycle, JSON-RPC handling,
auth, the developer-activity classifier, and tests. No public API change:
McpServer is re-exported from integrations::mcp so main.rs is unchanged.

Layout:
- mcp/mod.rs (13 LOC) — module declarations + McpServer re-export
- mcp/server.rs (92 LOC) — bind/serve/shutdown
- mcp/rpc.rs (330 LOC) — JSON-RPC dispatch, tool listing/calls, telemetry push
- mcp/auth.rs (32 LOC) — bearer-token check
- mcp/developer_activity/mod.rs (294 LOC) — policy-check -> activity projection
- mcp/developer_activity/shell_tokens.rs (305 LOC) — tokenizer + helpers
- mcp/developer_activity/package_command.rs (176 LOC) — npm/pnpm/pip/...
- mcp/developer_activity/cloud_command.rs (130 LOC) — aws/gcloud/gh/vercel/...
- mcp/developer_activity/secret_targets.rs (62 LOC) — secret-path detection
- mcp/developer_activity/metadata.rs (51 LOC) — activity IDs + metadata
- mcp/tests.rs (616 LOC) — all 18 tests moved unchanged

All production files are below the 600-line cap. The developer-activity
classifier stays in-tree per the swarm synthesis (deferred extraction).
