+++ apps/agent/src-tauri/src/integrations/mod.rs
@@ -3,9 +3,9 @@
 //! This module contains integrations for various AI coding tools.
 
 pub mod claude_code;
-pub mod mcp_server;
+pub mod mcp;
 pub mod openclaw_plugin;
 
 pub use claude_code::ClaudeCodeIntegration;
-pub use mcp_server::McpServer;
+pub use mcp::McpServer;
 pub use openclaw_plugin::OpenClawPluginIntegration;
